requirements.txt
torch>=2.1
torchvision
tqdm
pillow
scikit-learn
kaggle
